requirements.txt
bluepy
pycryptodomex
paho-mqtt
